Improve sync time.
